% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boost-tree.R
\name{adaboost}
\alias{adaboost}
\title{AdaBoost}
\usage{
adaboost(
  formula = NULL,
  data = NULL,
  epochs = 1000,
  tolerance = 1e-10,
  weak_learner = c("decision_stump", "perceptron"),
  x = NULL,
  y = NULL
)
}
\arguments{
\item{formula}{A formula.}

\item{data}{A data.frame.}

\item{epochs}{The maximum number of boosting iterations
to be run (0 will run until convergence.)}

\item{tolerance}{The tolerance for change in values of the weighted error during training.}

\item{weak_learner}{Weak learner to use.
Either "decision_stump" or "perceptron".}

\item{x}{Design matrix.}

\item{y}{Response matrix.}
}
\value{
An object of class \code{baritsu_ab}.
}
\description{
A wrapper around \code{\link[mlpack:adaboost]{mlpack::adaboost()}} that allows passing a formula.
}
\seealso{
\code{\link[mlpack:adaboost]{mlpack::adaboost()}} \code{\link[=predict.baritsu_ab]{predict.baritsu_ab()}}
}
