% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{package}
\name{baritsu-package}
\alias{baritsu}
\alias{baritsu-package}
\title{baritsu: Wrappers Around 'mlpack'}
\description{
A collection of wrappers around the 'mlpack' package that allows passing formula as their argument.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://paithiov909.github.io/baritsu/}
  \item Report bugs at \url{https://github.com/paithiov909/baritsu/issues}
}

}
\author{
\strong{Maintainer}: Akiru Kato \email{paithiov909@gmail.com}

}
\keyword{internal}
