% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision-tree.R
\name{decision_trees}
\alias{decision_trees}
\title{Decision trees}
\usage{
decision_trees(
  formula = NULL,
  data = NULL,
  tree_depth = 0,
  min_n = 20,
  minimum_gain_split = 1e-07,
  weights = NULL,
  x = NULL,
  y = NULL
)
}
\arguments{
\item{formula}{A formula.
Alternatively, a recipe object can be passed for this argument.
If a recipe is passed, \code{data} is ignored.}

\item{data}{A data.frame.}

\item{tree_depth}{Maximum depth of the tree.}

\item{min_n}{Minimum number of data points in a leaf.}

\item{minimum_gain_split}{Minimum gain required to split an internal node.}

\item{weights}{Weights for each observation.}

\item{x}{Design matrix.}

\item{y}{Response matrix.}
}
\value{
An object of class \code{baritsu_dt}.
}
\description{
A wrapper around \code{\link[mlpack:decision_tree]{mlpack::decision_tree()}} that allows passing a formula.
}
\details{
To prevent masking of \code{\link[parsnip:decision_tree]{parsnip::decision_tree()}},
this function is named \code{decision_trees} (plural form!)
}
\seealso{
\code{\link[mlpack:decision_tree]{mlpack::decision_tree()}} \code{\link[=predict.baritsu_dt]{predict.baritsu_dt()}}
}
