% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boost-tree-baritsu.R
\name{details_boost_tree_baritsu}
\alias{details_boost_tree_baritsu}
\title{AdaBoost via baritsu}
\description{
\code{\link[mlpack:adaboost]{mlpack::adaboost()}} is an implementation of
the AdaBoost.MH (Adaptive Boosting) algorithm for classification.
}
\details{
For this engine, there is a single mode: classification
\subsection{Tuning Parameters}{

This model has no tuning parameters.
}
}
\keyword{internal}
