% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lars-baritsu.R
\name{lars}
\alias{lars}
\alias{details_lars_baritsu}
\title{Least angle regression}
\usage{
lars(
  mode = "regression",
  engine = "baritsu",
  penalty_L1 = NULL,
  penalty_L2 = NULL
)
}
\arguments{
\item{mode}{A single character string for the type of model.
The only possible value for this model is "regression".}

\item{engine}{A single character string specifying what computational engine
to use for fitting.}

\item{penalty_L1}{Regularization parameter for L1-norm penalty.}

\item{penalty_L2}{Regularization parameter for L2-norm penalty.}
}
\description{
\code{lars()} defines a model that can predict numeric values from
predictors using \code{\link[=linear_regression]{linear_regression()}}, a wrapper of \code{\link[mlpack:lars]{mlpack::lars()}}.

\code{\link[mlpack:lars]{mlpack::lars()}} is an implementation of Least Angle Regression
(Stagewise/laSso), also known as LARS.
}
\details{
For this model, there is a single mode: regression
\subsection{Tuning Parameters}{

This model has 2 tuning parameters:
\itemize{
\item \code{penalty_L1} Amount of regularization for Lasso Penalty (type: double)
\item \code{penalty_L2} Amount of regularization for Ridge Penalty (type: double)
}
}
}
\seealso{
\code{\link[=linear_regression]{linear_regression()}}
}
\keyword{internal}
