% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-reg-baritsu.R
\name{details_linear_reg_baritsu}
\alias{details_linear_reg_baritsu}
\alias{register_linear_reg_baritsu}
\title{Linear regression via baritsu}
\usage{
register_linear_reg_baritsu()
}
\description{
\code{\link[mlpack:lars]{mlpack::lars()}} is an implementation of Least Angle Regression
(Stagewise/laSso), also known as LARS.
}
\details{
For this engine, there is a single mode: regression
\subsection{Tuning Parameters}{

This model has 1 tuning parameters:
\itemize{
\item \code{mixture} Proportion of Lasso Penalty (type: double, default: 1.0)
}

A value of \code{mixture = 1} corresponds to a pure lasso model,
while \code{mixture = 0} indicates ridge regression.
}

\subsection{Preprocessing requirements}{

Factor/categorical predictors need to be converted to numeric values
(e.g., dummy or indicator variables) for this engine.
When using the formula method via \code{\link[=fit.model_spec]{fit()}},
parsnip will convert factor columns to indicators.
}
}
\keyword{internal}
