% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic-reg-baritsu.R
\name{details_logistic_reg_baritsu}
\alias{details_logistic_reg_baritsu}
\title{Logistic regression via baritsu}
\description{
\code{\link[mlpack:logistic_regression]{mlpack::logistic_regression()}} is an implementation of
L2-regularized logistic regression for two-class classification.
}
\details{
For this engine, there is a single mode: classification
\subsection{Tuning Parameters}{

This model has 1 tuning parameter:
\itemize{
\item \code{penalty}: Amount of regularization (type: double)
}
}
}
\keyword{internal}
