% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlp-baritsu.R
\name{details_mlp_baritsu}
\alias{details_mlp_baritsu}
\alias{register_mlp_baritsu}
\title{Single layer neural network via baritsu}
\usage{
register_mlp_baritsu()
}
\description{
\code{\link[mlpack:perceptron]{mlpack::perceptron()}} is an implementation of a perceptron,
which is a single level neural network.
}
\details{
For this engine, there is a single mode: classification
\subsection{Tuning Parameters}{

This model has 1 tuning parameter:
\itemize{
\item \code{epochs}: Maximum number of iterations (type: integer)
}
}
}
\keyword{internal}
