% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinom-reg-baritsu.R
\name{details_multinom_reg_baritsu}
\alias{details_multinom_reg_baritsu}
\title{Multinomial regression via baritsu}
\description{
\code{\link[mlpack:softmax_regression]{mlpack::softmax_regression()}} is an implementation of
a softmax regression, a generalization of logistic regression
to the multiclass case, which has support for L2 regularization.
}
\details{
For this engine, there is a single mode: classification
\subsection{Tuning Parameters}{

This model has 1 tuning parameter:
\itemize{
\item \code{penalty}: Amount of Regularization (type: double)
}
}

\subsection{Preprocessing requirements}{

Predictors should have the same scale.
One way to achieve this is to center and scale each
so that each predictor has mean zero and a variance of one.
}
}
\keyword{internal}
