% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive-Bayes-baritsu.R
\name{details_naive_Bayes_baritsu}
\alias{details_naive_Bayes_baritsu}
\title{Naive Bayes models via baritsu}
\description{
\code{\link[mlpack:nbc]{mlpack::nbc()}} is an implementation of
a parametric Naive Bayes Classifier, used for classification.
}
\details{
For this engine, there is a single mode: classification
\subsection{Tuning Parameters}{

This model has no tuning parameters.
}
}
\keyword{internal}
