% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand-forest-baritsu.R
\name{details_rand_forest_baritsu}
\alias{details_rand_forest_baritsu}
\alias{register_rand_forest_baritsu}
\title{Random forests via baritsu}
\usage{
register_rand_forest_baritsu()
}
\description{
\code{\link[mlpack:random_forest]{mlpack::random_forest()}} is an implementation of
the standard random forest algorithm
by Leo Breiman for classification.
}
\details{
For this engine, there is a single mode: classification
\subsection{Tuning Parameters}{

This model has 3 tuning parameters:
\itemize{
\item \code{mtry}: Randomly Selected Predictors (type: integer)
\item \code{trees}: Trees (type: integer)
\item \code{min_n}: Minimal Node Size (type: integer)
}
}

\subsection{Preprocessing requirements}{

Factor/categorical predictors need to be converted to numeric values
(e.g., dummy or indicator variables) for this engine.
When using the formula method via \code{\link[=fit.model_spec]{fit()}},
parsnip will convert factor columns to indicators.
}
}
\keyword{internal}
