% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm-linear-baritsu.R
\name{details_svm_linear_baritsu}
\alias{details_svm_linear_baritsu}
\title{L2-regularized support vector machine via baritsu}
\description{
\code{\link[mlpack:linear_svm]{mlpack::linear_svm()}} is an implementation of
linear SVM for multiclass classification.
}
\details{
For this engine, there is a single mode: classification
\subsection{Tuning Parameters}{

This engine has 1 tuning parameter:
\itemize{
\item \code{margin}: Insensitivity Margin (type: double)
}
}

\subsection{Preprocessing requirements}{

Predictors should be centered and scaled.
One way to achieve this is to center and scale each
so that each predictor has mean zero and a variance of one.
}
}
\keyword{internal}
