% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-reg-bayesian.R
\name{linear_regression_bayesian}
\alias{linear_regression_bayesian}
\title{Bayesian linear regression}
\usage{
linear_regression_bayesian(
  formula = NULL,
  data = NULL,
  center = FALSE,
  scale = FALSE,
  x = NULL,
  y = NULL
)
}
\arguments{
\item{formula}{A formula.
Alternatively, a recipe object can be passed for this argument.
If a recipe is passed, \code{data} is ignored.}

\item{data}{A data.frame.}

\item{center}{Logical;
if enabled, centers the data and fit the intercept.}

\item{scale}{Logical;
if enabled, scales each feature by their standard deviations.}

\item{x}{Design matrix.}

\item{y}{Response matrix.}
}
\value{
An object of class \code{baritsu_blr}.
}
\description{
A wrapper around \code{\link[mlpack:bayesian_linear_regression]{mlpack::bayesian_linear_regression()}} that allows
passing a formula.
}
\seealso{
\code{\link[mlpack:bayesian_linear_regression]{mlpack::bayesian_linear_regression()}}
}
