% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic-reg.R
\name{logistic_regression}
\alias{logistic_regression}
\title{L2-regularized logistic regression}
\usage{
logistic_regression(
  formula = NULL,
  data = NULL,
  penalty = 1e-04,
  epochs = 1000,
  decision_boundary = 0.5,
  tolerance = 1e-10,
  optimizer = c("lbfgs", "sgd"),
  batch_size = 64,
  learn_rate = 0.01,
  x = NULL,
  y = NULL
)
}
\arguments{
\item{formula}{A formula.}

\item{data}{A data.frame.}

\item{penalty}{L2-regularization constant.}

\item{epochs}{Maximum number of iterations.}

\item{decision_boundary}{Decision boundary for prediction;
if the logistic function for a point is less than the boundary,
the class is taken to be 0; otherwise, the class is 1.}

\item{tolerance}{Convergence tolerance for optimizer.}

\item{optimizer}{Optimizer to use for training ("lbfgs" or "sgd").}

\item{batch_size}{Batch size for SGD.}

\item{learn_rate}{Step size for SGD optimizer.}

\item{x}{Design matrix.}

\item{y}{Response matrix.}
}
\value{
An object of class \code{baritsu_lgr}.
}
\description{
A wrapper around \code{\link[mlpack:logistic_regression]{mlpack::logistic_regression()}}
that allows passing a formula.
}
\seealso{
\code{\link[mlpack:logistic_regression]{mlpack::logistic_regression()}} \code{\link[=predict.baritsu_lgr]{predict.baritsu_lgr()}}
}
