% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive-bayes.R
\name{naive_bayes}
\alias{naive_bayes}
\title{Parametric naive Bayes classifier}
\usage{
naive_bayes(
  formula = NULL,
  data = NULL,
  incremental_variance = FALSE,
  x = NULL,
  y = NULL
)
}
\arguments{
\item{formula}{A formula.}

\item{data}{A data.frame.}

\item{incremental_variance}{Logical; passed to \code{\link[mlpack:nbc]{mlpack::nbc()}}.}

\item{x}{Design matrix.}

\item{y}{Response matrix.}
}
\value{
An object of class \code{baritsu_nbc}.
}
\description{
A wrapper around \code{\link[mlpack:nbc]{mlpack::nbc()}} that allows passing a formula.
}
\seealso{
\code{\link[mlpack:nbc]{mlpack::nbc()}} \code{\link[=predict.baritsu_nbc]{predict.baritsu_nbc()}}
}
