% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlp.R
\name{perceptron}
\alias{perceptron}
\title{Single level neural network}
\usage{
perceptron(formula = NULL, data = NULL, epochs = 100, x = NULL, y = NULL)
}
\arguments{
\item{formula}{A formula.
Alternatively, a recipe object can be passed for this argument.
If a recipe is passed, \code{data} is ignored.}

\item{data}{A data.frame.}

\item{epochs}{Maximum number of iterations.}

\item{x}{Design matrix.}

\item{y}{Response matrix.}
}
\value{
An object of class \code{baritsu_prc}.
}
\description{
A wrapper around \code{\link[mlpack:perceptron]{mlpack::perceptron()}}
that allows passing a formula.
}
\seealso{
\code{\link[mlpack:perceptron]{mlpack::perceptron()}}
}
