% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_predict.R, R/decision-tree.R,
%   R/hoeffding-tree.R, R/linear-reg-bayesian.R, R/linear-reg.R,
%   R/logistic-reg.R, R/mlp.R, R/multinom-reg.R, R/naive-bayes.R,
%   R/rand-forest.R, R/svm-linear.R
\name{predict.baritsu}
\alias{predict.baritsu}
\alias{predict.baritsu_dt}
\alias{predict.baritsu_ht}
\alias{predict.baritsu_blr}
\alias{predict.baritsu_lr}
\alias{predict.baritsu_lgr}
\alias{predict.baritsu_prc}
\alias{predict.baritsu_sr}
\alias{predict.baritsu_nbc}
\alias{predict.baritsu_rf}
\alias{predict.baritsu_svm}
\title{Prediction using mlpack via baritsu}
\usage{
\method{predict}{baritsu_dt}(object, newdata, type = c("both", "class", "prob"), ...)

\method{predict}{baritsu_ht}(object, newdata, ...)

\method{predict}{baritsu_blr}(object, newdata, ...)

\method{predict}{baritsu_lr}(object, newdata, ...)

\method{predict}{baritsu_lgr}(object, newdata, type = c("both", "class", "prob"), ...)

\method{predict}{baritsu_prc}(object, newdata, ...)

\method{predict}{baritsu_sr}(object, newdata, type = c("both", "class", "prob"), ...)

\method{predict}{baritsu_nbc}(object, newdata, type = c("both", "class", "prob"), ...)

\method{predict}{baritsu_rf}(object, newdata, type = c("both", "class", "prob"), ...)

\method{predict}{baritsu_svm}(object, newdata, type = c("both", "class", "prob"), ...)
}
\arguments{
\item{object}{An object out of baritsu function.}

\item{newdata}{A data.frame.}

\item{type}{Type of prediction. One of "both", "class", or "prob".}

\item{...}{Not used.}
}
\value{
A tibble that contains the predictions and/or probabilities
(and also the standard deviations of the predictive distribution
only for \code{predict.baritsu_blr}).
}
\description{
Predicts with new data using a stored mlpack model.
}
