% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-reg-bayesian.R
\name{predict.baritsu_blr}
\alias{predict.baritsu_blr}
\title{Bayesian linear regression prediction}
\usage{
\method{predict}{baritsu_blr}(object, newdata, ...)
}
\arguments{
\item{object}{An object of class \code{baritsu_blr}.}

\item{newdata}{A data.frame.}

\item{...}{Not used.}
}
\value{
A tibble that contains the predicted values
and the standard deviations of the predictive distribution.
}
\description{
Predicts with new data using a Bayesian linear regression model.
}
