% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision-tree.R
\name{predict.baritsu_dt}
\alias{predict.baritsu_dt}
\title{Decision trees prediction}
\usage{
\method{predict}{baritsu_dt}(object, newdata, type = c("both", "class", "prob"), ...)
}
\arguments{
\item{object}{An object of class \code{baritsu_dt}.}

\item{newdata}{A data.frame.}

\item{type}{Type of prediction. One of "both", "class", or "prob".}

\item{...}{Not used.}
}
\value{
A tibble that contains the predicted values and/or probabilities.
}
\description{
Predicts with new data using a decision tree model.
}
