% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-reg.R
\name{predict.baritsu_lr}
\alias{predict.baritsu_lr}
\title{Linear regression prediction}
\usage{
\method{predict}{baritsu_lr}(object, newdata, ...)
}
\arguments{
\item{object}{An object of class \code{baritsu_lr}.}

\item{newdata}{A data.frame.}

\item{...}{Not used.}
}
\value{
A tibble that contains the predicted values.
}
\description{
Predicts with new data using a linear regression model.
}
