% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive-Bayes.R
\name{predict.baritsu_nbc}
\alias{predict.baritsu_nbc}
\title{Naive Bayes Classifier prediction}
\usage{
\method{predict}{baritsu_nbc}(object, newdata, type = c("both", "class", "prob"))
}
\arguments{
\item{object}{An object of class \code{baritsu_nbc}.}

\item{newdata}{A data.frame.}

\item{type}{Type of prediction. One of "both", "class", or "prob".}
}
\value{
A tibble that contains the predicted values and/or probabilities.
}
\description{
Predicts with new data using a Naive Bayes Classifier model.
}
