% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlp.R
\name{predict.baritsu_prc}
\alias{predict.baritsu_prc}
\title{Perceptron prediction}
\usage{
\method{predict}{baritsu_prc}(object, newdata, ...)
}
\arguments{
\item{object}{An object of class \code{baritsu_prc}.}

\item{newdata}{A data.frame.}

\item{...}{Not used.}
}
\value{
A tibble that contains predicted class.
}
\description{
Predicts with new data using a perceptron model.
}
