% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand-forest.R
\name{predict.baritsu_rf}
\alias{predict.baritsu_rf}
\title{Predict for rand_forest}
\usage{
\method{predict}{baritsu_rf}(object, newdata, type = c("both", "class", "prob"))
}
\description{
Predict for rand_forest
}
