% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinom-reg.R
\name{predict.baritsu_sr}
\alias{predict.baritsu_sr}
\title{Softmax regression prediction}
\usage{
\method{predict}{baritsu_sr}(object, newdata, type = c("both", "class", "prob"), ...)
}
\arguments{
\item{object}{An object of class \code{baritsu_sr}.}

\item{newdata}{A data.frame.}

\item{type}{Type of prediction. One of "both", "class", or "prob".}

\item{...}{Not used.}
}
\value{
A tibble that contains the predicted values and/or probabilities.
}
\description{
Predicts with new data using a softmax regression model.
}
