% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand-forest.R
\name{random_forest}
\alias{random_forest}
\title{Random forests}
\usage{
random_forest(
  formula = NULL,
  data = NULL,
  mtry = 0,
  trees = 10,
  min_n = 1,
  maximum_depth = 0,
  minimum_gain_split = 0,
  seed = 0,
  x = NULL,
  y = NULL
)
}
\arguments{
\item{formula}{A formula.
Alternatively, a recipe object can be passed for this argument.
If a recipe is passed, \code{data} is ignored.}

\item{data}{A data.frame.}

\item{mtry}{Subspace dimension.
If 0, autoselects the square root of data dimensionality.}

\item{trees}{Number of trees.}

\item{min_n}{Minimum number of data points in a leaf.}

\item{maximum_depth}{Maximum depth of the tree.}

\item{minimum_gain_split}{Minimum gain required to split an internal node.}

\item{seed}{Random seed. If 0, \code{std::time(NULL)} is used internally.}

\item{x}{Design matrix.}

\item{y}{Response matrix.}
}
\value{
An object of class \code{baritsu_rf}.
}
\description{
A wrapper around \code{\link[mlpack:random_forest]{mlpack::random_forest()}} that allows passing a formula.
}
\seealso{
\code{\link[mlpack:random_forest]{mlpack::random_forest()}} \code{\link[=predict.baritsu_rf]{predict.baritsu_rf()}}
}
