% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinom-reg.R
\name{softmax_regression}
\alias{softmax_regression}
\title{Softmax regression}
\usage{
softmax_regression(
  formula = NULL,
  data = NULL,
  penalty = 0.001,
  stop_iter = 400,
  no_intercept = FALSE,
  x = NULL,
  y = NULL
)
}
\arguments{
\item{formula}{A formula.
Alternatively, a recipe object can be passed for this argument.
If a recipe is passed, \code{data} is ignored.}

\item{data}{A data.frame.}

\item{penalty}{L2-regularization constant.}

\item{stop_iter}{Maximum number of iterations.}

\item{no_intercept}{Logical; passed to \code{\link[mlpack:softmax_regression]{mlpack::softmax_regression()}}.}

\item{x}{Design matrix.}

\item{y}{Response matrix.}
}
\value{
An object of class \code{baritsu_sr}.
}
\description{
A wrapper around \code{\link[mlpack:softmax_regression]{mlpack::softmax_regression()}}
that allows passing a formula.
}
\seealso{
\code{\link[mlpack:softmax_regression]{mlpack::softmax_regression()}}
}
