% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{package}
\name{baritsu-package}
\alias{baritsu}
\alias{baritsu-package}
\title{baritsu: Wrapper Around 'mlpack'}
\description{
A wrapper around the 'mlpack' package that enables to use formula interface for that functions.
}
\author{
\strong{Maintainer}: Akiru Kato \email{paithiov909@gmail.com}

}
\keyword{internal}
