% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision-tree-baritsu.R
\name{details_decision_tree_baritsu}
\alias{details_decision_tree_baritsu}
\alias{register_decision_tree_baritsu}
\title{Decision trees via baritsu}
\usage{
register_decision_tree_baritsu()
}
\description{
\code{\link[mlpack:decision_tree]{mlpack::decision_tree()}} is an implementation of
an ID3-style decision tree for classification,
which supports categorical data.
}
\details{
For this engine, there is a single mode: classification
\subsection{Tuning Parameters}{

This model has 2 tuning parameters:
\itemize{
\item \code{tree_depth}: Tree depth (type: integer)
\item \code{min_n}: Minimal Node Size (type: integer)
}
}

\subsection{Preprocessing requirements}{

Factor/categorical predictors need to be converted to numeric values
(e.g., dummy or indicator variables) for this engine.
When using the formula method via \code{\link[=fit.model_spec]{fit()}},
parsnip will convert factor columns to indicators.
}
}
\keyword{internal}
