% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand-forest-baritsu.R
\name{details_rand_forest_baritsu}
\alias{details_rand_forest_baritsu}
\title{Random forests via baritsu}
\description{
\code{\link[mlpack:random_forest]{mlpack::random_forest()}} is an implementation of
the standard random forest algorithm
by Leo Breiman for classification.
}
\details{
For this engine, there is a single mode: classification
\subsection{Tuning Parameters}{

This model has 3 tuning parameters:
\itemize{
\item \code{mtry}: Randomly Selected Predictors (type: integer)
\item \code{trees}: Trees (type: integer)
\item \code{min_n}: Minimal Node Size (type: integer)
}
}
}
\keyword{internal}
