% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-reg.R
\name{linear_regression}
\alias{linear_regression}
\title{Linear regression}
\usage{
linear_regression(
  formula = NULL,
  data = NULL,
  mixture = 1,
  no_intercept = FALSE,
  no_normalize = FALSE,
  use_cholesky = FALSE,
  x = NULL,
  y = NULL
)
}
\arguments{
\item{formula}{A formula.
Alternatively, a recipe object can be passed for this argument.
If a recipe is passed, \code{data} is ignored.}

\item{data}{A data.frame.}

\item{mixture}{Regularization parameter within \verb{[0, 1]} (default: 1.0)
This value will be passed to lambda1 (for l1-norm penalty) as is,
and \code{1.0 - mixture} will be passed to lambda2 (for l2-norm penalty).}

\item{no_intercept}{Logical; passed to \code{\link[mlpack:lars]{mlpack::lars()}}.}

\item{no_normalize}{Logical; passed to \code{\link[mlpack:lars]{mlpack::lars()}}.}

\item{use_cholesky}{Logical; passed to \code{\link[mlpack:lars]{mlpack::lars()}}.}

\item{x}{Design matrix.}

\item{y}{Response matrix.}
}
\value{
An object of class \code{baritsu_lr}.
}
\description{
A wrapper around \code{\link[mlpack:linear_regression]{mlpack::linear_regression()}} and \code{\link[mlpack:lars]{mlpack::lars()}}
that enables formula interface.
}
\details{
When the mixture is 0, this function
fallbacks to \code{\link[mlpack:linear_regression]{mlpack::linear_regression()}} for performance.
}
\seealso{
\code{\link[mlpack:linear_regression]{mlpack::linear_regression()}} \code{\link[mlpack:lars]{mlpack::lars()}}
}
