% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm-linear.R
\name{linear_svm}
\alias{linear_svm}
\title{L2-regularized support vector machine}
\usage{
linear_svm(
  formula = NULL,
  data = NULL,
  margin = 1,
  penalty = 1,
  stop_iter = 1000,
  no_intercept = FALSE,
  tolerance = 1e-10,
  optimizer = c("lbfgs", "psgd"),
  epochs = 50,
  shuffle = FALSE,
  step_size = 0.01,
  seed = 0,
  x = NULL,
  y = NULL
)
}
\arguments{
\item{formula}{A formula.
Alternatively, a recipe object can be passed for this argument.
If a recipe is passed, \code{data} is ignored.}

\item{data}{A data.frame.}

\item{margin}{Margin of difference between correct class and other classes.}

\item{penalty}{L2-regularization constant.}

\item{stop_iter}{Maximum iterations for optimizer (0 indicates no limit).}

\item{no_intercept}{Logical; passed to \code{\link[mlpack:linear_svm]{mlpack::linear_svm()}}.}

\item{tolerance}{Convergence tolerance for optimizer.}

\item{optimizer}{Optimizer to use for training ("lbfgs" or "psgd").}

\item{epochs}{Maximum number of full epochs over dataset for parallel SGD.}

\item{shuffle}{Logical; if true, doesn't shuffle the order
in which data points are visited for parallel SGD.}

\item{step_size}{Step size for parallel SGD optimizer.}

\item{seed}{Random seed. If 0, \code{std::time(NULL)} is used internally.}

\item{x}{Design matrix.}

\item{y}{Response matrix.}
}
\value{
An object of class \code{baritsu_svm}.
}
\description{
A wrapper around \code{\link[mlpack:linear_svm]{mlpack::linear_svm()}} that allows passing a formula.
}
\seealso{
\code{\link[mlpack:linear_svm]{mlpack::linear_svm()}} \code{\link[=predict.baritsu_svm]{predict.baritsu_svm()}}
}
