% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engines-baritsu.R
\name{parsnip_engines_baritsu}
\alias{parsnip_engines_baritsu}
\title{Parsnip engines in baritsu}
\description{
Parsnip engines in baritsu
}
\details{
baritsu has the following parsnip engines:
\itemize{
\item \link{details_boost_tree_baritsu}
\item \link{details_decision_tree_baritsu}
\item \link{details_hoeffding_tree_baritsu}
\item \link{details_logistic_reg_baritsu}
\item \link{details_mlp_baritsu}
\item \link{details_multinom_reg_baritsu}
\item \link{details_naive_Bayes_baritsu}
\item \link{details_rand_forest_baritsu}
\item \link{details_svm_linear_baritsu}
\item \link{details_lars_baritsu}
}
}
\keyword{internal}
