% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand-forest.R
\name{predict.baritsu_rf}
\alias{predict.baritsu_rf}
\title{Random forests prediction}
\usage{
\method{predict}{baritsu_rf}(object, newdata, type = c("both", "class", "prob"))
}
\arguments{
\item{object}{An object of class \code{baritsu_rf}.}

\item{newdata}{A data.frame.}

\item{type}{Type of prediction. One of "both", "class", or "prob".}
}
\value{
A tibble that contains the predicted values and/or probabilities.
}
\description{
Predicts with new data using a random forest model.
}
